+++ app/genesis.go
@@ -2,6 +2,9 @@ package app
 
 import (
 	"encoding/json"
+
+	"github.com/CosmWasm/wasmd/x/wasm"
+	wasmtypes "github.com/CosmWasm/wasmd/x/wasm/types"
 )
 
 // The genesis state of the blockchain is represented here as a map of raw json
@@ -16,5 +19,18 @@ type GenesisState map[string]json.RawMessage
 // NewDefaultGenesisState generates the default state for the application.
 func NewDefaultGenesisState() GenesisState {
 	encCfg := MakeEncodingConfig()
-	return ModuleBasics.DefaultGenesis(encCfg.Marshaler)
+	gen := ModuleBasics.DefaultGenesis(encCfg.Marshaler)
+
+	// here we override wasm config to make it permissioned by default
+	wasmGen := wasm.GenesisState{
+		Params: wasmtypes.Params{
+			CodeUploadAccess:             wasmtypes.AllowNobody,
+			InstantiateDefaultPermission: wasmtypes.AccessTypeEverybody,
+			// DefaultMaxWasmCodeSize limit max bytes read to prevent gzip bombs
+			// It is 1200 KB in x/wasm, update it later via governance if really needed
+			MaxWasmCodeSize: wasmtypes.DefaultMaxWasmCodeSize,
+		},
+	}
+	gen[wasm.ModuleName] = encCfg.Marshaler.MustMarshalJSON(&wasmGen)
+	return gen
 }