+++ app/upgrades/v7/constants.go
@@ -1,0 +1,3 @@
+package v7
+
+const UpgradeName = "v7"