+++ simapp/sim_test.go
@@ -78,6 +78,8 @@ func fullAppSimulation(tb testing.TB, is_testing bool) {
 		sdkSimapp.FlagPeriodValue,
 		app.MakeEncodingConfig(),
 		sdkSimapp.EmptyAppOptions{},
+		app.GetWasmEnabledProposals(),
+		app.EmptyWasmOpts,
 		interBlockCacheOpt(),
 		fauxMerkleModeOpt)
 
@@ -154,6 +156,8 @@ func TestAppStateDeterminism(t *testing.T) {
 				sdkSimapp.FlagPeriodValue,
 				app.MakeEncodingConfig(),
 				sdkSimapp.EmptyAppOptions{},
+				app.GetWasmEnabledProposals(),
+				app.EmptyWasmOpts,
 				interBlockCacheOpt())
 
 			fmt.Printf(